+++ src/ui/svelte/fancy-toolbar/utils.ts
@@ -7,8 +7,12 @@ export function matchIds(tb: ToolbarItem2, id: string): boolean {
   return tb.id === id;
 }
 
-export function styleToString(style: Record<string, any>): string {
-  return Object.entries(style)
+export function styleToString(style: Record<string, any>): string | null {
+  const entries = Object.entries(style);
+  if (!entries.length) {
+    return null;
+  }
+  return entries
     .filter(([, v]) => v !== null && v !== undefined)
     .map(([k, v]) => {
       const cssKey = k.replace(/([A-Z])/g, "-$1").toLowerCase();